Reorganize tests so they can be run from Ant and the IDE
