+++ FlagMaker/nbproject/project.properties
@@ -39,13 +39,16 @@ javac.classpath=\
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
+javac.external.vm=false
 javac.processorpath=\
     ${javac.classpath}
 javac.source=1.8
 javac.target=1.8
 javac.test.classpath=\
     ${javac.classpath}:\
-    ${build.classes.dir}
+    ${build.classes.dir}:\
+    ${libs.junit_4.classpath}:\
+    ${libs.hamcrest.classpath}
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
@@ -123,3 +126,4 @@ run.test.classpath=\
     ${build.test.classes.dir}
 source.encoding=UTF-8
 src.dir=src
+test.test.dir=test
