+++ FlagMaker/nbproject/project.xml
@@ -17,7 +17,9 @@
             <source-roots>
                 <root id="src.dir"/>
             </source-roots>
-            <test-roots/>
+            <test-roots>
+                <root id="test.test.dir" name="Test Packages"/>
+            </test-roots>
         </data>
     </configuration>
 </project>
